@@ -1,8 +1,11 @@
-use super::*;
+use super::{
+	ord_periodicity, Aperiodic, Constraint, Formatter, Info, Limbo, Metadata, Periodic, Result, Runnable, State,
+	PRIORITY,
+};
 use crate::kernel::time;
-use cmos::RTCDateTime;
+use cmos_rtc::Time;
 use core::{
-	cmp::Ordering::{self, *},
+	cmp::Ordering::{self, Equal, Greater, Less},
 	fmt::Debug,
 	time::Duration,
 };
@@ -17,6 +20,7 @@ pub struct Task {
 impl Task {
 	// constructor
 
+	#[must_use]
 	pub fn new(constraint: Constraint, code: Runnable) -> Task {
 		use time::get_datetime;
 
@@ -26,32 +30,42 @@ impl Task {
 	// accessors
 
 	#[inline]
+	#[must_use]
 	pub fn get_metadata(&self) -> &Metadata { &self.metadata }
 
+	#[must_use]
 	#[inline]
 	pub fn get_constraint(&self) -> &Constraint { &self.metadata.0 }
 
 	#[inline]
+	#[must_use]
 	pub fn get_periodicity(&self) -> &Option<Either<Periodic, Aperiodic>> { &(self.metadata.0).0 }
 
 	#[inline]
+	#[must_use]
 	pub fn get_runnable(&self) -> &Runnable { &self.code }
 
 	#[inline]
+	#[must_use]
 	pub fn get_priority(&self) -> PRIORITY { (self.metadata.0).1 }
 
+	#[must_use]
 	#[inline]
 	pub fn get_info(&self) -> &Info { &self.metadata.1 }
 
 	#[inline]
+	#[must_use]
 	pub fn get_state(&self) -> State { (self.metadata.1).0 }
 
 	#[inline]
+	#[must_use]
 	pub fn get_running_time(&self) -> &Duration { &(self.metadata.1).1 }
 
 	#[inline]
-	pub fn get_creation_time(&self) -> &RTCDateTime { &(self.metadata.1).2 }
+	#[must_use]
+	pub fn get_creation_time(&self) -> &Time { &(self.metadata.1).2 }
 
+	#[must_use]
 	pub fn get_estimated_remaining_time(&self) -> Option<Duration> {
 		match self.get_periodicity() {
 			Some(periodicity) => {
@@ -69,7 +83,7 @@ impl Task {
 	#[inline]
 	pub fn set_state(&mut self, state: State) { (self.metadata.1).0 = state; }
 
-	pub fn set_last_execution(&mut self, datetime: &RTCDateTime) -> bool {
+	pub fn set_last_execution(&mut self, datetime: &Time) -> bool {
 		if self.is_periodic() {
 			self.get_periodicity().unwrap().left().unwrap().2 = *datetime;
 			true
@@ -84,12 +98,15 @@ impl Task {
 	// other
 
 	#[inline]
+	#[must_use]
 	pub fn is_realtime(&self) -> bool { (self.metadata.0).0.is_some() }
 
 	#[inline]
+	#[must_use]
 	pub fn is_periodic(&self) -> bool { self.is_realtime() && (self.metadata.0).0.as_ref().unwrap().is_left() }
 
 	#[inline]
+	#[must_use]
 	pub fn is_aperiodic(&self) -> bool { self.is_realtime() && (self.metadata.0).0.as_ref().unwrap().is_right() }
 }
 
@@ -108,13 +125,13 @@ impl Debug for Task {
 
 impl PartialEq for Task {
 	#[inline]
-	fn eq(&self, other: &Self) -> bool { self as *const _ == other as *const _ }
+	fn eq(&self, other: &Self) -> bool { core::ptr::eq(self, other) }
 }
 
 impl Ord for Task {
 	fn cmp(&self, other: &Self) -> Ordering {
 		match (self.get_periodicity(), other.get_periodicity()) {
-			(Some(periodicity_a), Some(periodicity_b)) => ord_periodicity(&periodicity_a, &periodicity_b),
+			(Some(periodicity_a), Some(periodicity_b)) => ord_periodicity(periodicity_a, periodicity_b),
 			(Some(_), None) => Greater,
 			(None, Some(_)) => Less,
 			(None, None) => Equal,
