@@ -1,10 +1,10 @@
 use super::{queue_size, BLOCKED_QUEUE, READY_QUEUE};
-use arraydeque::ArrayDeque;
+use alloc::collections::VecDeque;
 use spin::Mutex;
 
-/// Reorder READY_QUEUE with the provided strategy if READY_QUEUE is not empty.
+/// Reorder `READY_QUEUE` with the provided strategy if `READY_QUEUE` is not empty.
 /// Return the data collected by `global_info()`.
-pub fn update(strategy: &Fn(&Mutex<ArrayDeque<[u8; 256]>>) -> Option<u8>) -> (u8, usize, usize, Option<u8>) {
+pub fn update<T: Fn(&Mutex<VecDeque<u8>>) -> Option<u8>>(strategy: T) -> (u8, usize, usize, Option<u8>) {
 	if 0 < queue_size(&READY_QUEUE) {
 		terminator(&READY_QUEUE);
 		strategy(&READY_QUEUE);
@@ -20,9 +20,10 @@ pub fn update(strategy: &Fn(&Mutex<ArrayDeque<[u8; 256]>>) -> Option<u8>) -> (u8
 /// Get scheduler's components info.
 /// Return a tuple containing :
 /// - the total number of processes
-/// - the number of processes in BLOCKED_QUEUE
-/// - the number of processes in READY_QUEUE
+/// - the number of processes in `BLOCKED_QUEUE`
+/// - the number of processes in `READY_QUEUE`
 /// - the PID of eventual running process
+#[must_use]
 pub fn global_info() -> (u8, usize, usize, Option<u8>) {
 	use super::{get_process_count, swapper::get_running};
 
@@ -30,7 +31,7 @@ pub fn global_info() -> (u8, usize, usize, Option<u8>) {
 }
 
 /// Remove the processes those deadlines have been missed in the given queue.
-fn terminator(queue: &Mutex<ArrayDeque<[u8; 256]>>) -> u8 {
+fn terminator(queue: &Mutex<VecDeque<u8>>) -> u8 {
 	use crate::kernel::{
 		scheduler::get_process_periodicity,
 		time::{dt_add_du, get_datetime},
@@ -40,7 +41,7 @@ fn terminator(queue: &Mutex<ArrayDeque<[u8; 256]>>) -> u8 {
 	let mut guard = queue.lock();
 	let mut counter = 0;
 	for index in 0..guard.len() {
-		if let Some(periodicity) = get_process_periodicity(index as u8) {
+		if let Some(periodicity) = get_process_periodicity(u8::try_from(index).unwrap()) {
 			match periodicity {
 				Left(periodic) => {
 					if dt_add_du(periodic.2, periodic.1).unwrap() < get_datetime() {
@@ -66,20 +67,20 @@ pub mod strategy {
 		process::ord_periodicity,
 		scheduler::{get_process_periodicity, get_process_priority, queue_front},
 	};
-	use arraydeque::ArrayDeque;
-	use core::cmp::Ordering::*;
+	use alloc::collections::VecDeque;
+	use core::cmp::Ordering::{Equal, Greater, Less};
 	use spin::Mutex;
 
-	/// Put the processes in READY_QUEUE by order of PID.
+	/// Put the processes in `READY_QUEUE` by order of PID.
 	/// Return the PID of the first process in the ready queue if it exists.
-	pub fn process_id(queue: &Mutex<ArrayDeque<[u8; 256]>>) -> Option<u8> {
+	pub fn process_id(queue: &Mutex<VecDeque<u8>>) -> Option<u8> {
 		queue.lock().as_mut_slices().0.sort_unstable();
 		queue_front(queue)
 	}
 
-	/// Put the processes in READY_QUEUE by order of priority.
+	/// Put the processes in `READY_QUEUE` by order of priority.
 	/// Return the PID of the first process in the ready queue if it exists.
-	pub fn priority(queue: &Mutex<ArrayDeque<[u8; 256]>>) -> Option<u8> {
+	pub fn priority(queue: &Mutex<VecDeque<u8>>) -> Option<u8> {
 		queue.lock().as_mut_slices().0.sort_unstable_by(|a, b| {
 			match (get_process_priority(*a), get_process_priority(*b)) {
 				(Some(a_priority), Some(b_priority)) => a_priority.cmp(&b_priority),
@@ -92,11 +93,11 @@ pub mod strategy {
 		queue_front(queue)
 	}
 
-	/// Put the processes in READY_QUEUE by order of deadline and priority.
+	/// Put the processes in `READY_QUEUE` by order of deadline and priority.
 	/// LOW can preempt MEDIUM, MEDIUM can preempt HIGH.
 	/// Return the PID of the first process in the ready queue if it exists.
-	pub fn modified_earliest_deadline_first(queue: &Mutex<ArrayDeque<[u8; 256]>>) -> Option<u8> {
-		use crate::kernel::process::PRIORITY::*;
+	pub fn modified_earliest_deadline_first(queue: &Mutex<VecDeque<u8>>) -> Option<u8> {
+		use crate::kernel::process::PRIORITY::{HIGH, LOW};
 
 		queue.lock().as_mut_slices().0.sort_unstable_by(|a, b| {
 			match (get_process_periodicity(*a), get_process_periodicity(*b)) {
@@ -122,9 +123,9 @@ pub mod strategy {
 		queue_front(queue)
 	}
 
-	/// Put the processes in READY_QUEUE by order of deadline and priority.
+	/// Put the processes in `READY_QUEUE` by order of deadline and priority.
 	/// Return the PID of the first process in the ready queue if it exists.
-	pub fn earliest_deadline_first(queue: &Mutex<ArrayDeque<[u8; 256]>>) -> Option<u8> {
+	pub fn earliest_deadline_first(queue: &Mutex<VecDeque<u8>>) -> Option<u8> {
 		queue.lock().as_mut_slices().0.sort_unstable_by(|a, b| {
 			match (get_process_periodicity(*a), get_process_periodicity(*b)) {
 				(Some(periodicity_a), Some(periodicity_b)) => ord_periodicity(&periodicity_a, &periodicity_b),
