@@ -10,15 +10,14 @@ pub struct Executor {
 }
 
 impl Executor {
+	#[must_use]
 	pub fn new() -> Self {
 		Executor { tasks: BTreeMap::new(), task_queue: Arc::new(ArrayQueue::new(100)), waker_cache: BTreeMap::new() }
 	}
 
 	pub fn spawn(&mut self, task: Task) {
 		let task_id = task.id;
-		if self.tasks.insert(task.id, task).is_some() {
-			panic!("task with same ID already in tasks");
-		}
+		assert!(self.tasks.insert(task.id, task).is_none(), "task with same ID already in tasks");
 		self.task_queue.push(task_id).expect("queue full");
 	}
 
@@ -26,13 +25,11 @@ impl Executor {
 		// destructure `self` to avoid borrow checker errors
 		let Self { tasks, task_queue, waker_cache } = self;
 
-		while let Ok(task_id) = task_queue.pop() {
-			let task = match tasks.get_mut(&task_id) {
-				Some(task) => task,
-				None => continue, // task no longer exists
-			};
+		while let Some(task_id) = task_queue.pop() {
+			let Some(task) = tasks.get_mut(&task_id) else { continue };
 			let waker = waker_cache.entry(task_id).or_insert_with(|| TaskWaker::new(task_id, task_queue.clone()));
 			let mut context = Context::from_waker(waker);
+
 			match task.poll(&mut context) {
 				Poll::Ready(()) => {
 					// task done -> remove it and its cached waker
