@@ -11,6 +11,7 @@ pub struct BumpAllocator {
 
 impl BumpAllocator {
 	/// Creates a new empty bump allocator.
+	#[must_use]
 	pub const fn new() -> Self { BumpAllocator { heap_start: 0, heap_end: 0, next: 0, allocations: 0 } }
 
 	/// Initializes the bump allocator with the given heap bounds.
@@ -31,10 +32,7 @@ unsafe impl GlobalAlloc for Locked<BumpAllocator> {
 		let mut bump = self.lock(); // get a mutable reference
 
 		let alloc_start = align_up(bump.next, layout.align());
-		let alloc_end = match alloc_start.checked_add(layout.size()) {
-			Some(end) => end,
-			None => return ptr::null_mut(),
-		};
+		let Some(alloc_end) = alloc_start.checked_add(layout.size()) else { return ptr::null_mut() };
 
 		if alloc_end > bump.heap_end {
 			ptr::null_mut() // out of memory
