@@ -1,10 +1,11 @@
-use bootloader::bootinfo::{MemoryMap, MemoryRegionType::Usable};
+use bootloader_api::info::{MemoryRegionKind::Usable, MemoryRegions};
 use x86_64::{
-	structures::paging::{FrameAllocator, OffsetPageTable, PageTable, PhysFrame, Size4KiB},
+	registers,
+	structures::paging::{FrameAllocator, Mapper, OffsetPageTable, Page, PageTable, PhysFrame, Size4KiB},
 	PhysAddr, VirtAddr,
 };
 
-/// Initialize a new OffsetPageTable.
+/// Initialize a new [`OffsetPageTable`].
 ///
 /// # Safety
 ///
@@ -12,9 +13,9 @@ use x86_64::{
 /// complete physical memory is mapped to virtual memory at the passed
 /// `physical_memory_offset`. Also, this function must be only called once
 /// to avoid aliasing `&mut` references (which is undefined behavior).
+#[must_use]
 pub unsafe fn init(physical_memory_offset: VirtAddr) -> OffsetPageTable<'static> {
-	let level_4_table = active_level_4_table(physical_memory_offset);
-	OffsetPageTable::new(level_4_table, physical_memory_offset)
+	OffsetPageTable::new(active_level_4_table(physical_memory_offset), physical_memory_offset)
 }
 
 /// Returns a mutable reference to the active level 4 table.
@@ -24,9 +25,7 @@ pub unsafe fn init(physical_memory_offset: VirtAddr) -> OffsetPageTable<'static>
 /// `physical_memory_offset`. Also, this function must be only called once
 /// to avoid aliasing `&mut` references (which is undefined behavior).
 unsafe fn active_level_4_table(physical_memory_offset: VirtAddr) -> &'static mut PageTable {
-	use x86_64::registers::control::Cr3;
-
-	let (level_4_table_frame, _) = Cr3::read();
+	let (level_4_table_frame, _) = registers::control::Cr3::read();
 
 	let phys = level_4_table_frame.start_address();
 	let virt = physical_memory_offset + phys.as_u64();
@@ -35,35 +34,51 @@ unsafe fn active_level_4_table(physical_memory_offset: VirtAddr) -> &'static mut
 	&mut *page_table_ptr // unsafe
 }
 
-/// A FrameAllocator that always returns `None`.
+/// Creates an example mapping for the given page to frame `0xb8000`.
+pub fn create_example_mapping(
+	page: Page,
+	mapper: &mut OffsetPageTable,
+	frame_allocator: &mut impl FrameAllocator<Size4KiB>,
+) {
+	use x86_64::structures::paging::PageTableFlags as Flags;
+
+	let frame = PhysFrame::containing_address(PhysAddr::new(0xb8000));
+	let flags = Flags::PRESENT | Flags::WRITABLE;
+
+	let map_to_result = unsafe { mapper.map_to(page, frame, flags, frame_allocator) };
+	map_to_result.expect("map_to failed").flush();
+}
+
+/// A [`FrameAllocator`] that always returns `None`.
 pub struct EmptyFrameAllocator;
 
 unsafe impl FrameAllocator<Size4KiB> for EmptyFrameAllocator {
 	fn allocate_frame(&mut self) -> Option<PhysFrame> { None }
 }
 
-/// A FrameAllocator that returns usable frames from the bootloader's memory map.
+/// A [`FrameAllocator`] that returns usable frames from the bootloader's memory map.
 pub struct BootInfoFrameAllocator {
-	memory_map: &'static MemoryMap,
+	memory_map: &'static MemoryRegions,
 	next: usize,
 }
 
 impl BootInfoFrameAllocator {
-	/// Create a FrameAllocator from the passed memory map.
+	/// Create a [`FrameAllocator`] from the passed memory map.
 	///
 	/// # Safety
 	///
 	/// This function is unsafe because the caller must guarantee that the passed memory map is valid.
 	/// The main requirement is that all frames that are marked as `USABLE` in it are really unused.
-	pub unsafe fn init(memory_map: &'static MemoryMap) -> Self { BootInfoFrameAllocator { memory_map, next: 0 } }
+	#[must_use]
+	pub unsafe fn init(memory_map: &'static MemoryRegions) -> Self { BootInfoFrameAllocator { memory_map, next: 0 } }
 
 	/// Returns an iterator over the usable frames specified in the memory map.
 	fn usable_frames(&self) -> impl Iterator<Item = PhysFrame> {
 		// get usable regions from memory map
 		let regions = self.memory_map.iter();
-		let usable_regions = regions.filter(|r| r.region_type == Usable);
+		let usable_regions = regions.filter(|r| r.kind == Usable);
 		// map each region to its address range
-		let addr_ranges = usable_regions.map(|r| r.range.start_addr()..r.range.end_addr());
+		let addr_ranges = usable_regions.map(|r| r.start..r.end);
 		// transform to an iterator of frame start addresses
 		let frame_addresses = addr_ranges.flat_map(|r| r.step_by(4096));
 		// create `PhysFrame` types from the start addresses
