@@ -7,13 +7,19 @@
 extern crate alloc;
 
 use alloc::{boxed::Box, vec::Vec};
-use bootloader::{entry_point, BootInfo};
+use bootloader_api::{entry_point, BootInfo};
 use core::panic::PanicInfo;
 use dandelion::kernel::vmm::allocator::HEAP_SIZE;
 
-entry_point!(main);
+pub static BOOTLOADER_CONFIG: BootloaderConfig = {
+	let mut config = BootloaderConfig::new_default();
+	config.mappings.physical_memory = Some(Mapping::Dynamic);
+	config
+};
 
-fn main(boot_info: &'static BootInfo) -> ! {
+entry_point!(main, config = &BOOTLOADER_CONFIG);
+
+fn main(boot_info: &'static mut BootInfo) -> ! {
 	use dandelion::kernel::vmm::{
 		allocator,
 		memory::{self, BootInfoFrameAllocator},
@@ -21,9 +27,9 @@ fn main(boot_info: &'static BootInfo) -> ! {
 	use x86_64::VirtAddr;
 
 	dandelion::init();
-	let phys_mem_offset = VirtAddr::new(boot_info.physical_memory_offset);
+	let phys_mem_offset = VirtAddr::new(boot_info.physical_memory_offset.into_option().unwrap());
 	let mut mapper = unsafe { memory::init(phys_mem_offset) };
-	let mut frame_allocator = unsafe { BootInfoFrameAllocator::init(&boot_info.memory_map) };
+	let mut frame_allocator = unsafe { BootInfoFrameAllocator::init(&boot_info.memory_regions) };
 	allocator::init_heap(&mut mapper, &mut frame_allocator).expect("heap initialization failed");
 
 	test_main();
